Decentralize solution registration
This allows each file in src/solutions to register its own function and
config. It means that (a) main.rs is now almost empty, and (b) there is
no need for all the boilerplate in solutions/mod.rs.

There is now a macro `euler::solution!` that generates a small function
to test each problem solution. It uses the `linkme` crate to distribute
the registration over many files.

The only downside is that `cargo fmt` doesn't find the Rust code in the
solutions directory; to format everything, use rustfmt:

rustfmt src/solutions/*.rs